+++ Vortragsmanager/Views/VersammlungRednerMaximizedControl.xaml.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace Vortragsmanager.Views
+{
+    /// <summary>
+    /// Interaction logic for VersammlungRednerMaximizedControl.xaml
+    /// </summary>
+    public partial class VersammlungRednerMaximizedControl : UserControl
+    {
+        public VersammlungRednerMaximizedControl()
+        {
+            InitializeComponent();
+        }
+    }
+}
