+++ Vortragsmanager/Views/VersammlungSettingsViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using DevExpress.Mvvm;
 using DevExpress.Xpf.Core;
+using DevExpress.Xpf.LayoutControl;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -10,439 +11,6 @@ using Vortragsmanager.Models;
 
 namespace Vortragsmanager.Views
 {
-    public class ConregationViewModel : ViewModelBase
-    {
-        public ConregationViewModel(Conregation versammlung)
-        {
-            Versammlung = versammlung;
-            RednerListe = new SpeakersViewModelCollection(versammlung);
-            DeleteCommand = new DelegateCommand(Delete);
-            NewPersonCommand = new DelegateCommand(NewPerson);
-            CalculateDistanceCommand = new DelegateCommand(CalculateDistance);
-        }
-
-        public DelegateCommand DeleteCommand { get; private set; }
-
-        public DelegateCommand NewPersonCommand { get; private set; }
-
-        public DelegateCommand CalculateDistanceCommand { get; private set; }
-
-        private bool _deleted = false;
-
-        public void Delete()
-        {
-            if (ThemedMessageBox.Show("Versammlung löschen",
-                $"Soll die Versammlung {Versammlung.Name} mit allen {RednerListe.Count} Rednern gelöscht werden?" + Environment.NewLine +
-                "Alle vergangenen Einladungen werden durch 'unbekannt' ersetzt," + Environment.NewLine +
-                "alle zukünftigen Einladungen und Anfragen werden gelöscht!",
-                MessageBoxButton.YesNo,
-                MessageBoxImage.Warning) == MessageBoxResult.No)
-                return;
-
-            while (RednerListe.Count > 0)
-            {
-                RednerListe[0].RednerLöschen(true);
-                RednerListe.RemoveAt(0);
-            }
-            //Einladungen
-            var einladungen = Core.DataContainer.MeinPlan
-                .Where(x => x.Status == EventStatus.Zugesagt)
-                .Cast<Invitation>()
-                .Where(x => x.AnfrageVersammlung == Versammlung)
-                .ToList();
-            foreach (var einladung in einladungen)
-                Core.DataContainer.MeinPlan.Remove(einladung);
-
-            //Anfragen
-            var anfragen = Core.DataContainer.OffeneAnfragen.Where(x => x.Versammlung == Versammlung).ToList();
-            foreach (var anfrage in anfragen)
-                Core.DataContainer.OffeneAnfragen.Remove(anfrage);
-
-            //Externe Vorträge in dieser Versammlung
-            var externeE = Core.DataContainer.ExternerPlan.Where(x => x.Versammlung == Versammlung);
-            foreach (var outside in externeE)
-            {
-                outside.Versammlung = null;
-            }
-
-            Core.DataContainer.Versammlungen.Remove(Versammlung);
-            Sichtbarkeit = Visibility.Collapsed;
-            RaisePropertyChanged(nameof(Sichtbarkeit));
-        }
-
-        public void NewPerson()
-        {
-            var redner = Core.DataContainer.FindOrAddSpeaker("Neuer Redner", Versammlung);
-            var rednerModel = new SpeakerViewModel(redner);
-            RednerListe.Add(rednerModel);
-            rednerModel.Select();
-        }
-
-        public void CalculateDistance()
-        {
-            var start = Core.DataContainer.MeineVersammlung;
-            var end = Versammlung;
-            Entfernung = Core.GeoApi.GetDistance(start, end);
-        }
-
-        public int? Entfernung
-        {
-            get
-            {
-                return Versammlung.Entfernung;
-            }
-            set
-            {
-                if (value != null)
-                {
-                    Versammlung.Entfernung = (int)value;
-                }
-                RaisePropertyChanged();
-            }
-        }
-
-        public int Jahr1 { get; } = DateTime.Today.Year;
-
-        public int Jahr2 => Jahr1 + 1;
-
-        public int Jahr3 => Jahr1 + 2;
-
-        public string ZusammenkunftszeitJahr1
-        {
-            get
-            {
-                return Versammlung.GetZusammenkunftszeit(Jahr1);
-            }
-            set
-            {
-                if (value != ZusammenkunftszeitJahr1)
-                    Versammlung.SetZusammenkunftszeit(Jahr1, value);
-                RaisePropertyChanged(ZusammenkunftszeitJahr1);
-            }
-        }
-
-        public string ZusammenkunftszeitJahr2
-        {
-            get
-            {
-                return Versammlung.GetZusammenkunftszeit(Jahr2);
-            }
-            set
-            {
-                if ((value != ZusammenkunftszeitJahr1) && (value != Versammlung.GetZusammenkunftszeit(Jahr2)))
-                    Versammlung.SetZusammenkunftszeit(Jahr1, value);
-                RaisePropertyChanged(ZusammenkunftszeitJahr2);
-            }
-        }
-
-        public string ZusammenkunftszeitJahr3
-        {
-            get
-            {
-                return Versammlung.GetZusammenkunftszeit(Jahr3);
-            }
-            set
-            {
-                if ((value != ZusammenkunftszeitJahr2) && (value != Versammlung.GetZusammenkunftszeit(Jahr3)))
-                    Versammlung.SetZusammenkunftszeit(Jahr3, value);
-                RaisePropertyChanged(ZusammenkunftszeitJahr3);
-            }
-        }
-
-        public bool EigeneVersammlung
-        {
-            get
-            {
-                return (Core.DataContainer.MeineVersammlung == Versammlung);
-            }
-            set
-            {
-                if ((value == true) && (ThemedMessageBox.Show(
-                    Properties.Resources.Achtung,
-                    "Willst du diese Versammlung wirklich als deine eigene Versammlung setzen?",
-                    MessageBoxButton.YesNo,
-                    MessageBoxImage.Warning) == MessageBoxResult.Yes))
-                    Core.DataContainer.MeineVersammlung = Versammlung;
-            }
-        }
-
-        public Brush EigeneVersammlungTextInOrange
-        {
-            get
-            {
-                return (EigeneVersammlung) ? Brushes.Orange : Brushes.White;
-            }
-        }
-
-        public Conregation Versammlung { get; private set; }
-
-        public SpeakersViewModelCollection RednerListe { get; private set; }
-
-        public DevExpress.Xpf.LayoutControl.GroupBoxState IsSelected { get; set; }
-
-        public void Select(bool isSelected)
-        {
-            if (isSelected)
-            {
-                IsSelected = DevExpress.Xpf.LayoutControl.GroupBoxState.Maximized;
-            }
-            else
-            {
-                IsSelected = DevExpress.Xpf.LayoutControl.GroupBoxState.Normal;
-            }
-            RaisePropertyChanged(nameof(IsSelected));
-            RefreshVisibility();
-        }
-
-        private bool _matchFilter = true;
-
-        public bool MatchFilter
-        {
-            get
-            {
-                return _matchFilter;
-            }
-            set
-            {
-                _matchFilter = value;
-                RaisePropertyChanged();
-                RefreshVisibility();
-            }
-        }
-
-        private void RefreshVisibility()
-        {
-            if (_deleted)
-                Sichtbarkeit = Visibility.Collapsed;
-            else if (IsSelected == DevExpress.Xpf.LayoutControl.GroupBoxState.Maximized)
-                Sichtbarkeit = Visibility.Visible;
-            else if (EditMode)
-                Sichtbarkeit = Visibility.Collapsed;
-            else if (MatchFilter)
-                Sichtbarkeit = Visibility.Visible;
-            else
-                Sichtbarkeit = Visibility.Collapsed;
-        }
-
-        private Visibility _sichtbarkeit = Visibility.Visible;
-
-        public Visibility Sichtbarkeit
-        {
-            get
-            {
-                return _sichtbarkeit;
-            }
-            set
-            {
-                _sichtbarkeit = value;
-                RaisePropertyChanged();
-            }
-        }
-
-        private bool _editMode;
-
-        public bool EditMode
-        {
-            get
-            {
-                return _editMode;
-            }
-            set
-            {
-                _editMode = value;
-                RaisePropertyChanged();
-                RefreshVisibility();
-            }
-        }
-    }
-
-    public class SpeakerViewModel : ViewModelBase
-    {
-        public SpeakerViewModel(Speaker redner)
-        {
-            if (redner is null)
-                throw new NullReferenceException();
-
-            Redner = redner;
-            NeuerVortragCommand = new DelegateCommand(NeuenVortragSpeichern);
-            VortragLöschenCommand = new DelegateCommand(VortragLöschen);
-            DeleteCommand = new DelegateCommand(RednerLöschen);
-            _selectedVersammlung = Redner.Versammlung;
-        }
-
-        public DelegateCommand DeleteCommand { get; set; }
-
-        public DelegateCommand NeuerVortragCommand { get; private set; }
-
-        public DelegateCommand VortragLöschenCommand { get; private set; }
-
-        public void RednerLöschen(bool silent)
-        {
-            //ToDo: Redner löschen: im Vortragsplan die Zuteilungen ersetzen gegen "unbekannt"
-            if (silent || ThemedMessageBox.Show("Redner löschen",
-                $"Wirklich Redner {Redner.Name} löschen?" + Environment.NewLine +
-                "Alle vergangenen Einladungen werden durch 'unbekannt' ersetzt," + Environment.NewLine +
-                "alle zukünftigen Einladungen werden gelöscht.",
-                MessageBoxButton.YesNo,
-                MessageBoxImage.Warning) == MessageBoxResult.Yes)
-            {
-                //Einladungen
-                var einladungen = Core.DataContainer.MeinPlan
-                    .Where(x => x.Status == EventStatus.Zugesagt)
-                    .Cast<Invitation>()
-                    .Where(x => x.Ältester == Redner)
-                    .ToList();
-                foreach (var einladung in einladungen)
-                {
-                    if (einladung.Datum < DateTime.Today)
-                        einladung.Ältester = null;
-                    else
-                        Core.DataContainer.MeinPlan.Remove(einladung);
-                }
-                //Offene Anfragen
-                var anfragen = Core.DataContainer.OffeneAnfragen
-                    .Where(x => x.RednerVortrag.ContainsKey(Redner))
-                    .ToList();
-                foreach (var anfrage in anfragen)
-                {
-                    anfrage.RednerVortrag.Remove(Redner);
-                    if (anfrage.RednerVortrag.Count == 0)
-                        Core.DataContainer.OffeneAnfragen.Remove(anfrage);
-                }
-
-                //Externe Einladungen
-                if (Redner.Versammlung == Core.DataContainer.MeineVersammlung)
-                {
-                    var externeE = Core.DataContainer.ExternerPlan.Where(x => x.Ältester == Redner).ToList();
-                    foreach (var einladung in externeE)
-                        Core.DataContainer.ExternerPlan.Remove(einladung);
-                }
-
-                Core.DataContainer.Redner.Remove(Redner);
-                Sichtbar = Visibility.Collapsed;
-                RaisePropertyChanged(nameof(Sichtbar));
-            }
-        }
-
-        public void RednerLöschen()
-        {
-            RednerLöschen(false);
-        }
-
-        public Visibility Sichtbar { get; set; }
-
-        public void VortragLöschen()
-        {
-            Redner.Vorträge.Remove(GewählterVortrag);
-            RaisePropertyChanged(nameof(Vorträge));
-        }
-
-        public void NeuenVortragSpeichern()
-        {
-            if (string.IsNullOrWhiteSpace(NeueVorträgeListe))
-                NeueVorträgeListe = NeuerVortrag.Nummer.ToString(Core.DataContainer.German);
-
-            var nummern = NeueVorträgeListe.Split(new char[] { ',', ' ', ';' });
-            int num;
-            foreach (var nr in nummern)
-            {
-                bool isNum = int.TryParse(nr, out num);
-                var neuerV = Core.DataContainer.FindTalk(num);
-                if (neuerV == null)
-                    continue;
-                if (!Redner.Vorträge.Contains(neuerV))
-                    Redner.Vorträge.Add(neuerV);
-            }
-
-            RaisePropertyChanged(nameof(Vorträge));
-        }
-
-        public string NeueVorträgeListe
-        {
-            get; set;
-        }
-
-        public SolidColorBrush AktivBrush => Redner.Aktiv ? new SolidColorBrush(Colors.Green) : new SolidColorBrush(Colors.Red);
-
-        public string AktivText => Redner.Aktiv ? "Aktiv" : "Inaktiv";
-
-        public SolidColorBrush ÄltesterBrush => Redner.Ältester ? new SolidColorBrush(Colors.DodgerBlue) : new SolidColorBrush(Colors.Brown);
-
-        public string ÄltesterText => Redner.Ältester ? "Ältester" : "DAG";
-
-        public string Overview
-        {
-            get
-            {
-                string o = "Vorträge: ";
-                foreach (var v in Vorträge)
-                    o += v.Nummer + ", ";
-
-                return o.TrimEnd(' ').TrimEnd(',');
-            }
-        }
-
-        public Talk GewählterVortrag { get; set; }
-
-        public Talk NeuerVortrag { get; set; }
-
-        public ObservableCollection<Talk> Vorträge
-        {
-            get => new ObservableCollection<Talk>(Redner.Vorträge.OrderBy(x => x.Nummer));
-        }
-
-        public static ObservableCollection<Talk> Vortragsliste => Core.DataContainer.Vorträge;
-
-        public bool Ältester
-        {
-            get => Redner.Ältester;
-            set
-            {
-                Redner.Ältester = value;
-                RaisePropertiesChanged(new[] { "Ältester", "Typ" });
-            }
-        }
-
-        public DevExpress.Xpf.LayoutControl.GroupBoxState IsSelected { get; set; }
-
-        public void Select()
-        {
-            IsSelected = DevExpress.Xpf.LayoutControl.GroupBoxState.Maximized;
-            RaisePropertyChanged(nameof(IsSelected));
-        }
-
-        public Speaker Redner { get; private set; }
-
-        public ObservableCollection<Conregation> Versammlungen => new ObservableCollection<Conregation>(Core.DataContainer.Versammlungen);
-
-        public bool VersammlungenPopUp { get; set; }
-
-        private Conregation _selectedVersammlung;
-
-        public Conregation SelectedVersammlung
-        {
-            get
-            {
-                return _selectedVersammlung;
-            }
-            set
-            {
-                if (!VersammlungenPopUp && value != null && value != Redner?.Versammlung)
-                {
-                    if (ThemedMessageBox.Show("Achtung", $"Soll der Redner {Redner.Name} wirklich in die Versammlung {value.Name} verschoben werden?") == MessageBoxResult.OK)
-                    {
-                        Redner.Versammlung = value;
-                        _selectedVersammlung = value;
-                        Sichtbar = Visibility.Collapsed;
-                        RaisePropertyChanged(nameof(Sichtbar));
-                        //ToDo: UI der übernehmenden Versammlung aktualisieren
-                    }
-                }
-                RaisePropertyChanged();
-            }
-        }
-    }
-
     public class SpeakersViewModelCollection : ObservableCollection<SpeakerViewModel>
     {
         public SpeakersViewModelCollection(Conregation versammlung)
